==> 01_data_processing.ipynb <==
### Import libraries

import os
import pandas as pd
import re
import string

### Custom import functions

# scan root folder for filenames folder names (categories)
def get_filepaths_and_categories(root_folder):
    filepaths_and_categories = []
    for root, dirs, files in os.walk(root_folder):
        for file_name in files:
            directory = root[root.rfind('\\')+1 :]
            path_to_file = root + '/' + file_name
            filepaths_and_categories.append({'path': path_to_file, 
                                             'category': directory})
    return filepaths_and_categories

# import data from files and remove unnecessary headers
def import_data(filepaths_and_categories):
    texts = []
    categories = []

    for item in filepaths_and_categories:
        f = open(item['path'], mode='r', encoding="utf8", errors='ignore')
        lines = f.read().splitlines()

        def filters(line):
            if line == '': return False
            if re.match('(^[A-Z]).*(: )', line): return False
            if re.match('^(In article <)', line): return False
            else: return True

        lines_filtered = filter(filters, lines)

        texts.append(' '.join(lines_filtered))
        categories.append(item['category'])
    
    return {'corpus': texts, 'category': categories}

### Import data

#apply custom functions to import train data and convert to pandas DF
root_folder_train = 'data/20news-bydate-train'
train_files_and_categories = get_filepaths_and_categories(root_folder_train)
train_data = import_data(train_files_and_categories)

train_df = pd.DataFrame(train_data)

#...test data...
root_folder_test = 'data/20news-bydate-test'
test_files_and_categories = get_filepaths_and_categories(root_folder_test)
test_data = import_data(test_files_and_categories)

test_df = pd.DataFrame(test_data)

### Review data

#train data
train_df.head()

train_df.groupby(['category']).agg(['count'])

train_df.count()

#test data
test_df.head()

test_df.groupby(['category']).agg(['count'])

test_df.count()

# review random message from train dataset
train_df.at[8180, 'corpus']

### Data cleaning

# custom function for text cleaning
def clean_text(text):
    text = re.sub('(Subject: |Re: )', '', text) #1
    text = text.lower() #2
    # text = re.sub(r'\b[a-zA-Z]\b', "", text)
    text = re.sub('[%s]' % re.escape(string.punctuation), '', text)
    text = re.sub('(\t)', '', text)
    text = re.sub('\w*\d\w*', '', text)
    text = ' '.join(text.split())
    return text

# apply cleaning function
train_df['corpus'] = train_df['corpus'].apply(lambda x: clean_text(x))
test_df['corpus'] = test_df['corpus'].apply(lambda x: clean_text(x))

#Review message after cleaning
train_df.at[8180, 'corpus']

### Remove stopwords

# load libraries and stopwords disctionary
from nltk.corpus import stopwords
stopwords = stopwords.words('english')

# custom function to remove stopwords
def remove_stopwords(corpus):
    text = corpus.split()
    text = [word for word in text if not word in set(stopwords)]
    text = ' '.join(text)
    return text

# apply remove stopwords function
train_df['corpus'] = train_df['corpus'].apply(lambda x: remove_stopwords(x))
test_df['corpus'] = test_df['corpus'].apply(lambda x: remove_stopwords(x))

# review message after removing stopwords
train_df.at[8180, 'corpus']

### Lemmatization

from nltk.stem import WordNetLemmatizer
lemmatizer = WordNetLemmatizer()

# custom lemmatize function
def lemmatize(corpus):
    words = corpus.split()
    text = []
    for word in words:
        text.append(lemmatizer.lemmatize(word))
    text = ' '.join(text)
    return text

# apply lemmatize function
train_df['corpus'] = train_df['corpus'].apply(lambda x: lemmatize(x))
test_df['corpus'] = test_df['corpus'].apply(lambda x: lemmatize(x))

# review message after lemmatization
train_df.at[8180, 'corpus']

### Save train_df and test_df variables to pickles

train_df.to_pickle('pickles/train_df.pkl')
test_df.to_pickle('pickles/test_df.pkl')
