==> 02_NB_SVM_models.ipynb <==
### Load libraries

import pandas as pd
import numpy as np

### Import data from pickles

train_df = pd.read_pickle('pickles/train_df.pkl')
test_df = pd.read_pickle('pickles/test_df.pkl')

### Vectorization

# apply vectorization on train set
from sklearn.feature_extraction.text import CountVectorizer
cv = CountVectorizer()
texts_train = train_df['corpus'].tolist()
X_train = cv.fit_transform(texts_train).toarray()
y_train = train_df.iloc[:, -1].values

# ...test set
texts_test = test_df['corpus'].tolist()
X_test = cv.transform(texts_test).toarray()
y_test = test_df.iloc[:, -1].values

### Naive Bayes model

# import and train model
from sklearn.naive_bayes import MultinomialNB
classifier = MultinomialNB()
classifier.fit(X_train, y_train)

# predictions
y_pred = classifier.predict(X_test)

# accuracy score and confusion matrix
from sklearn.metrics import confusion_matrix, accuracy_score
print(accuracy_score(y_test, y_pred))
cm = confusion_matrix(y_test, y_pred, labels=['Automotive', 'Computers', 'Politics', 'Religion', 'Science', 'Sports'])
print(cm)

### Support Vector Machine model

# import and train model
from sklearn.svm import LinearSVC
classifier = LinearSVC(max_iter=2000)
classifier.fit(X_train, y_train)

#preditions
y_pred = classifier.predict(X_test)

# accuracy score and confusion matrix
from sklearn.metrics import confusion_matrix, accuracy_score
print(accuracy_score(y_test, y_pred))
cm = confusion_matrix(y_test, y_pred)
print(cm)
