==> 03_BERT_model.ipynb <==
### Load libraries

import os
import numpy as np
import tensorflow as tf
import ktrain
from ktrain import text
import pandas as pd

### Import data from pickles

train_df = pd.read_pickle('pickles/train_df.pkl')
test_df = pd.read_pickle('pickles/test_df.pkl')

### Preprocess train and test sets

(X_train, y_train), (X_test, y_test), preproc = text.texts_from_df(train_df=train_df,
                                                                   text_column = 'corpus',
                                                                   label_columns = ['category'],
                                                                   val_df = test_df,
                                                                   maxlen = 500,
                                                                   preprocess_mode = 'bert')

### Create model

model = text.text_classifier(name='bert', train_data=(X_train, y_train), preproc=preproc)

### Train model

learner = ktrain.get_learner(model=model, train_data=(X_train, y_train), val_data=(X_test, y_test), batch_size=2)

learner.fit_onecycle(lr=2e-5, epochs=5)

### Evaluate model

learner.evaluate((X_test, y_test), class_names=preproc.get_classes())
