==> 04_AWS_Comprehend.ipynb <==
## 1. Prepare training and test data

### Load libraries

import pandas as pd
import codecs

### Import data from pickles

train_df = pd.read_pickle('pickles/train_df.pkl')
test_df = pd.read_pickle('pickles/test_df.pkl')

# review samlple data
train_df.sample(n=5)

### Change columns order

col_name = 'category'

# training data
first_col = train_df.pop(col_name)
train_df.insert(0, col_name, first_col)
print(train_df.sample(n=5))

# test data
first_col = test_df.pop(col_name)
test_df.insert(0, col_name, first_col)
print(test_df.sample(n=5))

### Save training data into .csv file

train_df.to_csv("data/Comprehend/train.csv", index=False, header=False)

### Preparing .txt file with training data

# remove categories column
test_df.drop(columns='category', axis=0, inplace=True)

test_df.head()

# save file
with codecs.open('data/Comprehend/test.txt', 'w', "utf-8-sig") as f:
    for index, row in test_df.iterrows():
        f.write("%s\n" % row['corpus'])

## 2. Read and evaluate model predictions

### Load libraries

import pandas as pd
import json

### Open predictions file

with open("data/Comprehend/predictions.jsonl", "r") as file:
    json_list = list(file)

obj_list = []

for json_str in json_list:
    result = json.loads(json_str)
    obj_list.append(result)

print(obj_list[0])

### Prepare list with predicted classes (categories)

predicted_classes = []

for obj in obj_list:
    predicted_classes.append(obj['Classes'][0]['Name'])

print(predicted_classes[0])

### Load test data

test_df = pd.read_pickle('pickles/test_df.pkl')

### Read test categories

y_test = test_df.iloc[:, -1].values
print(y_test)

### Prepare accuracy score and confusion matrix

from sklearn.metrics import confusion_matrix, accuracy_score
print(accuracy_score(y_test, predicted_classes))
cm = confusion_matrix(y_test, predicted_classes, labels=['Automotive', 'Computers', 'Politics', 'Religion', 'Science', 'Sports'])
print(cm)
